Api_Class_Assignments/Saturday_Class_assignments/Class_08/Assignment_02/pyproject.toml
[project]
name = "assignment-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gemini-agent-helper>=0.1.1",
    "openai-agents>=0.2.8",
    "pydantic>=2.11.7",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]
